adding get low

+++ postscrape/postscrape/spiders/posts_spider.py
@@ -17,8 +17,12 @@ from statistics import mean
 
 class PostsSpider(scrapy.Spider):
     name = 'posts'
+    #######begin bagian yang diubah ubah
+    partisi = 0
+    #######end bagian yang diubah ubah
     start_urls = []
-    itung = 0
+    list_emiten_used = []
+    itung = 0    
     baris_analyst = 1
     baris = 1
     sebelum = datetime.now()
@@ -31,8 +35,8 @@ class PostsSpider(scrapy.Spider):
     def __init__(self):
         print('init dimulai')
         self.sebelum = datetime.now()
-        self.delete_files()
         self.isi_urls()
+        self.delete_files()
         self.excel_initialization()
 
     def parse(self, response):
@@ -52,51 +56,78 @@ class PostsSpider(scrapy.Spider):
 
         if self.itung == len(self.start_urls):
             self.wb_analyst.save('/home/prasojo/belajar/BelajarData/dataset/saham/JII_analysis.xlsx')
-            sleep(5) # menaruh sleep penting agar proses save document semua selesai total
             from matplotlib.pyplot import savefig
             import mplfinance as mpf
             directory = '/home/prasojo/belajar/BelajarData/dataset/saham/emiten'
             pic_directory = '/home/prasojo/belajar/BelajarData/dataset/saham/emiten_pic'
             pic_directory_up = '/home/prasojo/belajar/BelajarData/dataset/saham/emiten_pic_up'
-            count = 1
+            count_pic = 1
             stats_data = []
             for filename in os.listdir(directory):
                 file_name = f'{directory}/{filename}'
                 df_saham = pd.read_excel(file_name)
                 stat_data = self.stat_process(df_saham)
                 stats_data.append(stat_data)
-                if stat_data[2] == 'up_near':
-                    df_saham = df_saham.reindex(index=df_saham.index[::-1])
-                    df_saham = df_saham.set_index(pd.DatetimeIndex(df_saham['date'].values))
-                    emiten = filename.split('.')[0]
-                    print(emiten, count)
-                    pic_file_name = f'{pic_directory_up}/{emiten}.jpg'
-                    save = dict(fname=pic_file_name,dpi=100)
-                    mpf.plot(df_saham,type='candle',mav=(3,7),volume=True,figsize=(25,15),savefig=save)
-                    count += 1                    
-                elif df_saham['JII'][0] in self.list_jii70:
-                    df_saham = df_saham.reindex(index=df_saham.index[::-1])
-                    df_saham = df_saham.set_index(pd.DatetimeIndex(df_saham['date'].values))
-                    emiten = filename.split('.')[0]
-                    print(emiten, count)
-                    pic_file_name = f'{pic_directory}/{emiten}.jpg'
-                    save = dict(fname=pic_file_name,dpi=100)
-                    mpf.plot(df_saham,type='candle',mav=(3,7),volume=True,figsize=(25,15),savefig=save)
-                    count += 1
-            df_stat = pd.DataFrame(stats_data,columns=['emiten', 'macd', 'up_near_macd', 'rsi','percent_volum'])
+                if df_saham['JII'][0] in self.list_emiten_used:
+                    if stat_data[2] == 'up_near' or stat_data[1] == 'up' or stat_data[3] == 'low':
+                        df_saham = df_saham.reindex(index=df_saham.index[::-1])
+                        df_saham = df_saham.set_index(pd.DatetimeIndex(df_saham['date'].values))
+                        emiten = filename.split('.')[0]
+                        print(emiten, count_pic)
+                        pic_file_name = f'{pic_directory_up}/{emiten}.jpg'
+                        save = dict(fname=pic_file_name,dpi=100)
+                        mpf.plot(df_saham,type='candle',mav=(3,7),volume=True,figsize=(25,15),savefig=save)
+                        count_pic += 1                    
+                    elif df_saham['JII'][0] in self.list_jii70:
+                        df_saham = df_saham.reindex(index=df_saham.index[::-1])
+                        df_saham = df_saham.set_index(pd.DatetimeIndex(df_saham['date'].values))
+                        emiten = filename.split('.')[0]
+                        print(emiten, count_pic)
+                        pic_file_name = f'{pic_directory}/{emiten}.jpg'
+                        save = dict(fname=pic_file_name,dpi=100)
+                        mpf.plot(df_saham,type='candle',mav=(3,7),volume=True,figsize=(25,15),savefig=save)
+                        count_pic += 1
+            df_stat = pd.DataFrame(stats_data,
+                                    columns=['emiten', 'macd', 'up_near_macd','get_low', 'rsi','percent_volum',
+                                            'Open','High','Low','Close','Volume'])
             df_stat.to_excel(f'/home/prasojo/belajar/BelajarData/dataset/saham/analisa_saham_harian_v3.xlsx')
             self.sesudah = datetime.now()
             selisih = self.sesudah - self.sebelum
             print(selisih)
+            print(f'selesai pakai partisi {self.partisi}')
 
 
     def stat_process(self, df):
         emiten = df['JII'][0]
+        last_open = df['Open'][0]
+        last_high = df['High'][0]
+        last_low = df['Low'][0]
+        last_close = df['Close'][0]
+        last_volume = df['Volume'][0]
         macd = self.calculate_macd(df['Close'])
         up_near_macd = self.calculate_up_near_macd(df['Close'])
         rsi = self.calculate_rsi(df['Close'])
         percent_volum = self.calculate_percent_volum(df['Volume'])
-        return [emiten,macd,up_near_macd,rsi,percent_volum]
+        get_low = self.get_low(df)
+        print(df['date'][0])
+        return [emiten, macd, up_near_macd, get_low, rsi, percent_volum, last_open, last_high,
+        last_low,last_close,last_volume]
+
+    def get_low(self, df):
+        current_close = df['Close'][0]
+        previous_close = df['Close'][1]
+        # current_open = df['Open'][0]
+        # current_high = df['High'][0]
+        rsi5 = self.calculate_rsi(df['Close'], rsi_coverage=5)
+        con_a = current_close >= previous_close
+        # con_b = current_close > current_open
+        # con_c = current_close == current_open
+        # con_d = current_high > (1.01*current_close)
+        # con_e = con_b or (con_c and con_d)
+        con_f = rsi5 <= 30
+        if con_a and con_f:
+            return 'low'
+        return '---'
 
     def calculate_macd(self, df_close, ma_small=3, ma_big=7):
         current_ma_small = mean(df_close[:ma_small])
@@ -125,7 +156,7 @@ class PostsSpider(scrapy.Spider):
         current_ma_big = mean(df_close[:ma_big])
         previous_ma_big = mean(df_close[1:ma_big+1])
 
-        con_a = current_ma_big > current_ma_small
+        con_a = current_ma_big >= current_ma_small
         con_b = current_ma_small > previous_ma_small
         con_c = (current_ma_big-current_ma_small) < (previous_ma_big-previous_ma_small)
         con_d = previous_ma_big > previous_ma_small
@@ -155,8 +186,16 @@ class PostsSpider(scrapy.Spider):
         cup[cup<0]=0
         cdown[cdown>0]=0
         cdown = cdown * -1
-        up_value = np.mean(cup[:rsi_coverage])
-        down_value = np.mean(cdown[:rsi_coverage])
+        if max(cup[:rsi_coverage]) == 0:
+            up_value = 0.001
+        else:
+            up_value = np.mean(cup[:rsi_coverage])
+
+        if max(cdown[:rsi_coverage]) == 0:
+            down_value = 0.001
+        else:
+            down_value = np.mean(cdown[:rsi_coverage])
+
         banding = up_value/down_value
         rsi = 100 - (100/(banding+1))
         return rsi
@@ -175,7 +214,7 @@ class PostsSpider(scrapy.Spider):
             count = 0
             
             for data in kolom:
-                if len(kolom) < 6 :
+                if len(kolom) < 6 :                    
                     break
 
                 volume = kolom[6].text
@@ -210,6 +249,8 @@ class PostsSpider(scrapy.Spider):
                     value = datetime.strftime(value, '%Y-%m-%d') 
                     pd_data_kolom.append(value)
                     pd_data_kolom.append(emiten)
+                    if row == rows[0]:
+                        print(value, emiten)
 
                     if baris_emiten <= 30:
                         self.sheet_analyst.write(self.baris_analyst, 0, value) 
@@ -261,10 +302,15 @@ class PostsSpider(scrapy.Spider):
             self.list_jii70.append(emiten)
 
         sheet_jii_large = wb_jii_list.sheet_by_index(1)
-        # self.start_urls.append('https://finance.yahoo.com/quote/AALI.JK/history?p=AALI.JK')
+        # self.start_urls.append('https://finance.yahoo.com/quote/HRUM.JK/history?p=HRUM.JK')
+        # self.list_emiten_used.append('HRUM')
+        hitung = 0
         for i in range(sheet_jii_large.nrows):
             emiten = sheet_jii_large.cell_value(i, 0)
-            self.start_urls.append('https://finance.yahoo.com/quote/'+emiten+'.JK/history?p='+emiten+'.JK')
+            if (hitung%3) == self.partisi:
+                self.list_emiten_used.append(emiten)
+                self.start_urls.append('https://finance.yahoo.com/quote/'+emiten+'.JK/history?p='+emiten+'.JK')
+            hitung = hitung + 1
 
     def excel_initialization(self):
         self.sheet_analyst = self.wb_analyst.add_sheet('Sheet 1')
@@ -285,6 +331,8 @@ class PostsSpider(scrapy.Spider):
         for dir in list_dir:
             if len(os.listdir(dir)) > 0:
                 for filename in os.listdir(dir):
+                    emiten = filename.split('.')[0]
                     file_name = f'{dir}/{filename}'
                     if os.path.isfile(file_name):
-                        os.remove(file_name)
+                        if emiten in self.list_emiten_used:
+                            os.remove(file_name)
